+++ attributes/logstash_client_setup.rb
@@ -1,8 +1,5 @@
 
-default['logstash']['supervisor_gid'] = 'adm'
-default['lostash']['beaver']['outputs'] = [{
-  "rabbitmq" => {
-    "exchange_type" => "direct",
-    "exchange" => "logstash-exchange"
-  }
-}]
+default_unless['elkstack']['config']['custom_logstash'] = {}
+default['elkstack']['config']['agent_protocol'] = 'lumberjack'
+default['elkstack']['config']['cloud_monitoring']['enabled'] = false
+default['rsyslog']['port'] = '5959'