+++ recipes/logstash_nginx_modifications.rb
@@ -1,0 +1,3 @@
+
+execute "chmod 755 /var/log/nginx"
+execute "chmod 644 $(find /var/log/nginx -type f)"