+++ recipes/sensu_server_setup.rb
@@ -28,7 +28,7 @@ node.set['uchiwa']['api'] = [
 ]
 
 include_recipe "uchiwa"
-include_recipe "nginx_ssl_setup" if node['roles'].include?('https')
+include_recipe "TheCheftacularCookbook::nginx_ssl_setup" if node['roles'].include?('https')
 include_recipe "nginx"
 
 include_recipe "TheCheftacularCookbook::sensu_gems"
@@ -84,3 +84,5 @@ include_recipe "TheCheftacularCookbook::sensu_server_cron_setup"
 include_recipe "sensu::server_service"
 
 include_recipe "sensu::api_service"
+
+execute "chmod 755 /var/log/sensu"